Relocate helper test classes
Move helper/utility test classes to "neutral" package (instead of being
in packages of specific add-ons), so that they can be used by all
add-ons without depending on each other, also, add JavaDoc to some
classes and methods.
Move common code useful for several test types to base class (TestUtils)
having other specific classes extend it (ActiveScannerTestUtils and
PassiveScannerTestUtils).
Parameterise PassiveScannerTestUtils and update tests accordingly.
