+++ test/org/zaproxy/zap/extension/ascanrules/TestParameterTamperUnitTest.java
@@ -27,6 +27,7 @@ import static org.junit.Assert.assertThat;
 
 import org.junit.Test;
 import org.parosproxy.paros.core.scanner.Alert;
+import org.zaproxy.zap.testutils.NanoServerHandler;
 
 import fi.iki.elonen.NanoHTTPD;
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
