+++ test/org/zaproxy/zap/extension/ascanrules/TestSQLInjectionUnitTest.java
@@ -29,6 +29,7 @@ import org.junit.Test;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.zaproxy.zap.model.Tech;
 import org.zaproxy.zap.model.TechSet;
+import org.zaproxy.zap.testutils.NanoServerHandler;
 
 import fi.iki.elonen.NanoHTTPD;
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
